player moves

+++ includes/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: dbouron <dbouron@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/22 11:05:01 by dbouron           #+#    #+#             */
-/*   Updated: 2022/10/07 12:01:37 by dbouron          ###   ########lyon.fr   */
+/*   Updated: 2022/10/07 13:36:25 by dbouron          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -101,8 +101,8 @@ typedef struct s_player
 {
 	int		x;
 	int		y;
-	int		x_dir;
-	int		y_dir;
+	double	x_dir;
+	double	y_dir;
 	double	angle;
 	int		dist_from_proj_plane;
 	int		height;
